Added base activities
