+++ app/src/main/java/com/blstream/studybox/activities/base/BaseBasicActivity.java
@@ -1,33 +1,21 @@
-package com.blstream.studybox.activities;
+package com.blstream.studybox.activities.base;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.design.widget.TextInputEditText;
+import android.support.v7.app.AppCompatActivity;
 
 import com.blstream.studybox.ConnectionStatusReceiver;
-import com.blstream.studybox.auth.BaseAuthView;
-import com.hannesdorfmann.mosby.mvp.MvpBasePresenter;
-import com.hannesdorfmann.mosby.mvp.viewstate.MvpViewStateActivity;
 
 import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper;
 
-public abstract class BaseAuthActivity<V extends BaseAuthView, P extends MvpBasePresenter<V>>
-        extends MvpViewStateActivity<V, P> {
-
-    protected final static float ENABLED_BUTTON_ALPHA = 1.0f;
-    protected final static float DISABLED_BUTTON_ALPHA = 0.5f;
+public abstract class BaseBasicActivity extends AppCompatActivity {
 
     protected ConnectionStatusReceiver connectionStatusReceiver;
 
-    protected abstract void setError(String message);
-    protected abstract void setFieldError(TextInputEditText field, String message);
-    protected abstract void setFormEnabled(boolean enabled);
-
     @Override
-    public void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         connectionStatusReceiver = new ConnectionStatusReceiver();
-        setRetainInstance(true);
     }
 
     @Override
